#3 Diagonal "clusters" filled
Made a quite confusing algorithm for populating the diagonal "clusters" of the Sudoku board. This will probably be changed in future but it works for now. Clusters are being filled because they are independent with respect to each other, meaning the backtracking alghorithm can solve them and it can always produce a random "shuffled" board.